code decomposition implemented for project components
