+++ src/canister_testing/test_canister1/Cargo.toml
@@ -8,7 +8,7 @@ candid = { version = "0.10" }
 ic-cdk = "0.14.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-evm_logs_types = { path = "../../evm_logs_types" } 
+evm_logs_types = { path = "../../../evm_logs_types" } 
 ic-cdk-macros = "0.9"
 ic-web3-rs = "0.1.9"
 hex = "0.4.3"
