+++ src/chain_service/logs_fetcher.rs
@@ -1,0 +1,35 @@
+use evm_rpc_canister_types::{
+    BlockTag, GetLogsArgs, GetLogsResult, MultiGetLogsResult, RpcServices, EvmRpcCanister, LogEntry,
+};
+use candid::Nat;
+
+pub async fn fetch_logs(
+    evm_rpc: &EvmRpcCanister,
+    rpc_providers: &RpcServices,
+    from_block: u64,
+    addresses: Option<Vec<String>>,
+    topics: Option<Vec<Vec<String>>>,
+) -> Result<Vec<LogEntry>, String> {
+    let get_logs_args = GetLogsArgs {
+        fromBlock: Some(BlockTag::Number(Nat::from(from_block))),
+        toBlock: Some(BlockTag::Latest),
+        addresses: addresses.unwrap_or_default(),
+        topics,
+    };
+
+    let cycles = 10_000_000_000;
+    let (result,) = evm_rpc
+        .eth_get_logs(rpc_providers.clone(), None, get_logs_args, cycles)
+        .await
+        .map_err(|e| format!("Call failed: {:?}", e))?;
+
+    match result {
+        MultiGetLogsResult::Consistent(r) => match r {
+            GetLogsResult::Ok(logs) => Ok(logs),
+            GetLogsResult::Err(err) => Err(format!("{:?}", err)),
+        },
+        MultiGetLogsResult::Inconsistent(_) => {
+            Err("RPC providers gave inconsistent results".to_string())
+        }
+    }
+}
