+++ src/chain_service/mod.rs
@@ -1,0 +1,8 @@
+pub mod config;
+pub mod service;
+pub mod monitoring;
+pub mod logs_fetcher;
+pub mod events_processor;
+pub mod utils;
+
+pub use config::ChainConfig;
